Devise forms

+++ db/seeds.rb
@@ -8,13 +8,11 @@
 require 'open-uri'
 
 puts "deleting data"
-User.destroy_all
-Offer.destroy_all
+Deal.destroy_all if Rails.env.development?
+Offer.destroy_all if Rails.env.development?
 puts "DB Clean!"
 
 puts "Seeding..."
-user = User.new(email: "luca@test.com", first_name: "Luca", last_name: "Test", password: "123456")
-user.save!
 
 offer_1 = Offer.new(
   instrument: "guitar" ,
